services/api/requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.29.0,<0.31.0
SQLAlchemy>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
pydantic>=2.6.0,<3.0.0
PyJWT>=2.8.0,<3.0.0
alembic>=1.13.0,<2.0.0
